requirements.txt
flask
flask-smorest
flask-sqlalchemy
passlib
marshmallow
uuid
python-dotenv
gunicorns